Added Location Service and icons

+++ app/src/main/java/com/yelloyew/ewaweather/data/network/ForecastService.kt
@@ -11,8 +11,8 @@ interface ForecastService {
     @Headers("X-Yandex-API-Key: 64d1b01c-af63-4f41-abb1-6cb31ad20be9")
     @GET("forecast?")
     suspend fun weatherSevenDays(
-        @Query("lat") latitude : Float = 55.75f,
-        @Query("lon") longitude : Float = 37.61f,
+        @Query("lat") latitude : Double = 55.75,
+        @Query("lon") longitude : Double = 37.61,
         @Query("lang") language : String = "ru",
         @Query("limit") count_days : Int = 7,
         @Query("hours") per_hour : Boolean = false
